Refactored request dtos to request folder
