+++ user-service/src/main/java/rs/banka4/user_service/dto/requests/CreateClientDto.java
@@ -1,15 +1,14 @@
-package rs.banka4.user_service.dto;
+package rs.banka4.user_service.dto.requests;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import rs.banka4.user_service.models.Privilege;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Set;
+
 @Schema(description = "DTO for creating Client")
 public record CreateClientDto(
         @Schema(description = "Client's first name", example = "Ognjen")
