+++ user-service/src/main/java/rs/banka4/user_service/service/abstraction/ClientService.java
@@ -4,9 +4,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import rs.banka4.user_service.dto.ClientDto;
-import rs.banka4.user_service.dto.CreateClientDto;
+import rs.banka4.user_service.dto.requests.CreateClientDto;
 import rs.banka4.user_service.dto.PrivilegesDto;
-import rs.banka4.user_service.dto.UpdateClientDto;
+import rs.banka4.user_service.dto.requests.UpdateClientDto;
 
 public interface ClientService {
     ResponseEntity<PrivilegesDto> getPrivileges(String token);
