+++ user-service/src/main/java/rs/banka4/user_service/service/abstraction/EmployeeAbs.java
@@ -1,6 +1,6 @@
 package rs.banka4.user_service.service.abstraction;
 
-import rs.banka4.user_service.dto.CreateEmployeeRequest;
+import rs.banka4.user_service.dto.requests.CreateEmployeeRequest;
 import rs.banka4.user_service.models.Employee;
 
 public interface EmployeeAbs {
