+++ user-service/src/main/java/rs/banka4/user_service/service/impl/ClientServiceImpl.java
@@ -5,10 +5,11 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import rs.banka4.user_service.dto.*;
+import rs.banka4.user_service.dto.requests.CreateClientDto;
+import rs.banka4.user_service.dto.requests.UpdateClientDto;
 import rs.banka4.user_service.exceptions.NotAuthenticated;
 import rs.banka4.user_service.exceptions.NotFound;
 import rs.banka4.user_service.mapper.BasicClientMapper;
