+++ user-service/src/main/java/rs/banka4/user_service/service/impl/EmployeeServiceImpl.java
@@ -13,6 +13,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import rs.banka4.user_service.config.RabbitMqConfig;
 import rs.banka4.user_service.dto.*;
+import rs.banka4.user_service.dto.requests.CreateEmployeeDto;
+import rs.banka4.user_service.dto.requests.UpdateEmployeeDto;
 import rs.banka4.user_service.exceptions.*;
 import rs.banka4.user_service.mapper.BasicEmployeeMapper;
 import rs.banka4.user_service.mapper.EmployeeMapper;
