+++ user-service/src/test/java/rs/banka4/user_service/generator/EmployeeObjectMother.java
@@ -1,6 +1,6 @@
 package rs.banka4.user_service.generator;
 
-import rs.banka4.user_service.dto.CreateEmployeeDto;
+import rs.banka4.user_service.dto.requests.CreateEmployeeDto;
 import rs.banka4.user_service.models.Privilege;
 
 import java.time.LocalDate;
